수정

+++ src/main/java/com/project/Instagram/domain/member/service/RefreshTokenService.java
@@ -2,7 +2,6 @@ package com.project.Instagram.domain.member.service;
 
 import com.project.Instagram.domain.member.entity.RefreshToken;
 import com.project.Instagram.global.error.BusinessException;
-import com.project.Instagram.global.error.ErrorCode;
 import com.project.Instagram.global.jwt.RefreshTokenRedisRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
@@ -10,6 +9,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+
+import static com.project.Instagram.global.error.ErrorCode.*;
 
 @Service
 @RequiredArgsConstructor
@@ -19,13 +20,11 @@ public class RefreshTokenService {
     @SneakyThrows
     @Transactional
     public void deleteRefreshTokenByValue(Long memberId) {
-        try {
             List<RefreshToken> refreshTokens = refreshTokenRedisRepository.findAllByMemberId(memberId);
-            for (RefreshToken refreshToken : refreshTokens)
+            if (refreshTokens.size() == 0)  throw new BusinessException(MEMBER_ID_REFRESH_TOKEN_DOES_NOT_EXIST);
+            for (RefreshToken refreshToken : refreshTokens) {
                 refreshTokenRedisRepository.delete(refreshToken);
-        } catch (BusinessException e) {
-            throw new BusinessException(ErrorCode.MEMBER_ID_REFRESH_TOKEN_DOES_NOT_EXIST);
-        }
+            }
     }
 
     @SneakyThrows
